+++ src/app/locales/en/components/space-automation/execution-options-form.js
@@ -1,0 +1,7 @@
+export default {
+  form: {
+    logLevel: {
+      label: 'Logging level',
+    },
+  },
+};