+++ src/app/locales/en/components/space-automation/run-workflow-creator.js
@@ -3,8 +3,10 @@ export default {
     list: {
       header: 'Choose workflow you would like to run',
     },
-    inputStores: {
-      header: 'Provide initial values for "{{atmWorkflowSchemaName}}" workflow',
+    setup: {
+      header: 'Run new "{{atmWorkflowSchemaName}}" workflow execution',
+      inputDataHeader: 'Initial values',
+      optionsHeader: 'Options',
       nothingToProvideMessage: 'This workflow does not need any initial values.',
       runWorkflowBtn: 'Run workflow',
       backBtn: 'Back',