+++ src/app/services/production/workflow-manager.js
@@ -178,18 +178,21 @@ export default Service.extend({
   },
 
   /**
-   * @param {String} atmWorkflowSchemaId
-   * @param {RevisionNumber} atmWorkflowSchemaRevisionNumber
-   * @param {String} spaceId
-   * @param {Object} storeInitialContentOverlay map (storeSchemaId => initial content)
+   * @param {String} options.atmWorkflowSchemaId
+   * @param {RevisionNumber} options.atmWorkflowSchemaRevisionNumber
+   * @param {String} options.spaceId
+   * @param {Object} options.storeInitialContentOverlay
+   *   map (storeSchemaId => initial content)
+   * @param {AuditLogEntrySeverity} options.logLevel
    * @returns {Promise<Models.AtmWorkflowExecution>}
    */
-  async runWorkflow(
+  async runWorkflow({
     atmWorkflowSchemaId,
     atmWorkflowSchemaRevisionNumber,
     spaceId,
-    storeInitialContentOverlay
-  ) {
+    storeInitialContentOverlay,
+    logLevel,
+  }) {
     return await this.get('store').createRecord('atmWorkflowExecution', {
       _meta: {
         additionalData: {
@@ -197,6 +200,7 @@ export default Service.extend({
           atmWorkflowSchemaRevisionNumber,
           spaceId,
           storeInitialContentOverlay,
+          logLevel,
         },
       },
     }).save();