+++ src/tests/integration/components/space-automation-test.js
@@ -12,6 +12,7 @@ import { getSlide } from '../../helpers/one-carousel';
 import sinon from 'sinon';
 import { suppressRejections } from '../../helpers/suppress-rejections';
 import { schedule } from '@ember/runloop';
+import { EntrySeverity } from 'onedata-gui-common/utils/audit-log';
 
 describe('Integration | Component | space-automation', function () {
   setupRenderingTest();
@@ -137,10 +138,16 @@ describe('Integration | Component | space-automation', function () {
     expect(createTabPane).to.have.class('active');
     expect(createTabPane.querySelector('.run-workflow-creator')).to.exist;
     await click(getSlide('list').querySelector('.revisions-table-revision-entry'));
-    await fillIn(getSlide('inputStores').querySelector('.form-control'), '10');
-    await click(getSlide('inputStores').querySelector('.btn-submit'));
+    await fillIn(getSlide('setup').querySelector('.form-control'), '10');
+    await click(getSlide('setup').querySelector('.btn-submit'));
     expect(runWorkflowStub).to.be.calledOnce
-      .and.to.be.calledWith('workflow1', 1, 'space1', sinon.match.any);
+      .and.to.be.calledWith({
+        atmWorkflowSchemaId: 'workflow1',
+        atmWorkflowSchemaRevisionNumber: 1,
+        spaceId: 'space1',
+        storeInitialContentOverlay: sinon.match.any,
+        logLevel: EntrySeverity.Info,
+      });
     expect(find('.nav-item-preview')).to.have.class('active');
   });